+++ lender/src/adapters/skip_while.rs
@@ -1,7 +1,8 @@
 use core::{fmt, ops::ControlFlow};
 
 use crate::{
-    try_trait_v2::Try, FallibleLend, FallibleLender, FallibleLending, FusedFallibleLender, FusedLender, Lend, Lender, Lending,
+    try_trait_v2::Try, FallibleLend, FallibleLender, FallibleLending, FusedFallibleLender, FusedLender, Lend, Lender,
+    Lending,
 };
 #[derive(Clone)]
 #[must_use = "lenders are lazy and do nothing unless consumed"]